percentage = (float)filesize_read / (float)filesize;

+++ ESP8266WirelessPrintAsync/ESP8266WirelessPrintAsync.ino
@@ -386,7 +386,7 @@ void setup() {
     // http://docs.octoprint.org/en/master/api/datamodel.html#sec-api-datamodel-jobs-job
     float percentage = 0.0;
     if( filesize_read > 0 ){
-      percentage = filesize_read / filesize; // Not super accurate but what OctoPrint does, too
+      percentage = (float)filesize_read / (float)filesize; // Not super accurate but what OctoPrint does, too
     }
     request->send(200, "application/json", "{\r\n  \"job\": {\r\n    \"file\": {\r\n      \"name\": \"Unknown\",\r\n      \"origin\": \"local\",\r\n      \"size\": " + String(filesize) + ",\r\n      \"date\": 1378847754\r\n    },\r\n    \"estimatedPrintTime\": 8811,\r\n    \"filament\": {\r\n      \"length\": 810,\r\n      \"volume\": 5.36\r\n    }\r\n  },\r\n  \"progress\": {\r\n    \"completion\": " + String(percentage) + ",\r\n    \"filepos\": " + String(filesize_read) + ",\r\n    \"printTime\": 0,\r\n    \"printTimeLeft\": 0\r\n  }\r\n}");
   });
